+++ pkg/remotewriteutils/util.go
@@ -41,5 +41,12 @@ func FetchPrometheusList(ctx context.Context, r *ResourceWrapper, rw *pmov1alpha
 		return nil, microerror.Maskf(errorFetchingPrometheus, "Could not fetch Prometheus with label selector %#q", rw.Spec.ClusterSelector.String())
 	}
 
+	for i, p := range prometheusList.Items {
+		if p.GetNamespace() == metav1.NamespaceSystem {
+			// Remove the prometheus-agent from the list.
+			prometheusList.Items = append(prometheusList.Items[:i], prometheusList.Items[i+1:]...)
+		}
+	}
+
 	return prometheusList, nil
 }