+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Ignore remotewrite feature in kube-system namespace.
+
 ## [4.17.0] - 2022-12-14
 
 ### Changed